[IMP] l10n_my_edi_extended: Malaysian fields in the portal
The POS supports to get your invoice using a QR code present
on a POS ticket after the fact.

Support for it is not easily done for public users and will
be supported later; but for logged in users we can display
the fields on their portal to allow them to fetch the invoice
correctly by themselves.

X-original-commit: 34ff862b7a634d2f94cc3124015b53497099ce49
Part-of: odoo/odoo#217604
Signed-off-by: Nicolas Viseur (vin) <vin@odoo.com>