[FIX] portal: identitycheck for internal users with passkey
When an internal user has a passkey and attempts to pass the portal
identitycheck, then the identitycheck's auth_method is set to webauthn
which breaks the flow since portal's identitycheck only supports
passwords.

This scenario is unlikely to occur but it does break the internal
users' portal. This is a temporary solution because once I implement
passkeys for portal users, I will allow portal users to use webauthn to
verify their identity.

closes odoo/odoo#219520

X-original-commit: 22c086fa7fb10b8424a6314a331e8864a9fbc5aa
Signed-off-by: Florian Vranckx (flvr) <flvr@odoo.com>
Signed-off-by: Bram Van Gaal (brvg) <brvg@odoo.com>