[IMP] account_peppol: Add a demo EAS for Peppol
When the company is registered with this particular EAS, Odoo will act
as if it was in demo mode.
No call to the real Peppol Network is performed, everything is mocked locally.
This allows to do the trainings on /trial databases, and get the lastests
improvements.

Also fixing some crons that needed to be run with the right company.

task-4968580

closes odoo/odoo#220705

X-original-commit: 66265b42e3e8e596e3fe2b0254c15173c3d4e0a3
Signed-off-by: Laurent Smet (las) <las@odoo.com>
Signed-off-by: Claire Bretton (clbr) <clbr@odoo.com>