[FIX] website: translate deepest nodes
Scenario:

- enable second language on website
- go to /shop/1 and try to translate description_ecommerce

Result: this is not translatable

Cause:

Since at least b455ea85853dfc19ed01e33986ad270cf80ee5d6 the
contenteditable attribute in ContentEditablePlugin is not set on an
element if it has a contenteditable ancestor.

TranslationPlugin disable all editable nodes containing editable nodes.

So with this combination, if we had a node for example:

```
<div class="oe_editable" data-oe-model="product.template" data-oe-id="1"
  data-oe-field="description_ecommerce" data-oe-type="html">
    <div>
        <span class="oe_editable" data-oe-model="product.template"
            data-oe-id="1" data-oe-field="description_ecommerce">
            test
        </span>
    </div>
</div>
```

the contenteditable was added to the parent div.oe_editable, but was
removed by TranslationPlugin so the "test" text was not translatable.

Fix: move the code that adds data-oe-readonly class in the
after_setup_editor_handlers so it is run before contenteditable
attributes are set.

opw-5128618

closes odoo/odoo#230885

X-original-commit: d4f8aebea8cc7eba1517575302e072e0d5e2e406
Signed-off-by: Colin Louis (loco) <loco@odoo.com>
Signed-off-by: Nicolas Lempereur (nle) <nle@odoo.com>