[FIX] product: fix discounted price for templates
Steps to reproduce:
1- Add a product that has variants to a vendor pricelist.
2- Set the vendor's unit for this product to a one
different than the product's unit.
3- Create an RFQ with from that vendor.
4- Open the catalog.

Issue:
`discounted_price` is miscalculated and causes wrong price calculation,
it shows the price multiplied by its factor.
i.e if a price of pack of 6 is 10$, it will be shown as 60$.

Cause:
We only check for `product_id` not for `product_tmpl_id` when
computing the `price_discounted`, so if its a template,
`product_id` is false, and its sent with `_compute_price()`
and the same price is returned since there's no unit and after that
its sent to `_get_product_price_and_data` where its converted to the
new unit.

closes odoo/odoo#240519

Signed-off-by: Quentin Wolfs (quwo) <quwo@odoo.com>